requirements.txt
SpeechRecognition>=3.14.5
PyAudio>=0.2.14
edge-tts>=7.2.7
pygame>=2.6.1
google-genai>=1.63.0